fix copy

+++ src/main.cpp
@@ -17,13 +17,13 @@ int main() {
     if (triangles_number < 0) {
         throw std::runtime_error("Input error: incorrect triangles_number");
     }
-    std::list<std::pair<Geometry::Triangle<float>, size_t>> triangles;
+    std::list<std::pair<Geometry::Triangle<double>, size_t>> triangles;
 
-    float x_min, y_min, z_min, x_max, y_max, z_max;
+    double x_min, y_min, z_min, x_max, y_max, z_max;
 
     for (int i = 0; i < triangles_number; i++) {
-        float x, y, z;
-        std::array<Geometry::Point<float>, 3> triangle_points;
+        double x, y, z;
+        std::array<Geometry::Point<double>, 3> triangle_points;
         for (int j = 0; j < 3; j++) {
             std::cin >> x >> y >> z;
             is_cin_good();
@@ -44,16 +44,16 @@ int main() {
             y_max = y > y_max ? y : y_max;
             z_max = z > z_max ? z : z_max;
         }
-        std::pair<Geometry::Triangle<float>, size_t> triangle =
-            std::pair<Geometry::Triangle<float>, size_t>(
-                Geometry::Triangle<float>(triangle_points[0],
+        std::pair<Geometry::Triangle<double>, size_t> triangle =
+            std::pair<Geometry::Triangle<double>, size_t>(
+                Geometry::Triangle<double>(triangle_points[0],
                                            triangle_points[1],
                                            triangle_points[2]),
                 i);
         triangles.push_back(triangle);
     }
 
-    Octree::Octree<float> tree = Octree::Octree<float>(
+    Octree::Octree<double> tree = Octree::Octree<double>(
         triangles,
         Octree::CubeParams{x_min, y_min, z_min, x_max, y_max, z_max});
